+++ core/Console/MakeMail.php
@@ -1,0 +1,99 @@
+<?php
+/**
+ * Create mailable instance
+ *
+ * @package Brocooly
+ * @subpackage Brocket
+ * @since 1.7.0
+ */
+
+declare(strict_types=1);
+
+namespace Brocooly\Console;
+
+use Brocooly\Mail\Mailable;
+use Symfony\Component\Console\Style\SymfonyStyle;
+use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
+
+class MakeMail extends CreateClassCommand
+{
+	/**
+	 * The name of the command
+	 *
+	 * @var string
+	 */
+	protected static $defaultName = 'new:mail';
+
+	/**
+	 * Generated class root namespace (its own namespace excluded)
+	 *
+	 * @var string
+	 */
+	protected string $rootNamespace = 'Theme\Mail';
+
+	/**
+	 * Under which path will be created file
+	 *
+	 * @var string
+	 */
+	protected string $themeFileFolder = 'Mail';
+
+	/**
+	 * @inheritDoc
+	 */
+	protected function configure(): void
+	{
+		$this
+			->addArgument(
+				'mail',
+				InputArgument::REQUIRED,
+				'Mailable name',
+			);
+	}
+
+	/**
+	 * @inheritDoc
+	 */
+	protected function execute( InputInterface $input, OutputInterface $output ) : int
+	{
+		$io = new SymfonyStyle( $input, $output );
+
+		$name = $input->getArgument( 'mail' );
+
+		$this->defineDataByArgument( $name );
+
+		$this->generateClassComments(
+			[
+				$this->className . " - custom theme mailable\n",
+			]
+		);
+
+		$class = $this->generateClassCap();
+
+		$this->createMethod( $class );
+		$this->createMethod( $class, 'build' );
+
+		$exists = $this->createFile( $this->file );
+		if ( $exists ) {
+			$io->warning( 'File ' . $exists . ' already exists' );
+			return CreateClassCommand::FAILURE;
+		}
+
+		$io->success( 'Mailable ' . $name . ' was successfully created' );
+		return CreateClassCommand::SUCCESS;
+	}
+
+	protected function generateClassCap()
+	{
+		$namespace = $this->file->addNamespace( $this->rootNamespace );
+		$namespace->addUse( Mailable::class );
+
+		$class = $namespace->addClass( $this->className );
+		$class->addExtend( Mailable::class );
+
+		return $class;
+	}
+
+}